Modified to handle h basis functions
Note that there is an error in the treatment of h function
normalization by both Psi4 and Gaussian. There is a workaround in the
current version of gdma.f90. If in the future the error in Psi4 or
Gaussian is corrected, change the value of the constant h_bug_fixed to
.true.